relocating RealtimeListener
